General cleanup:
Solution Cleanup and grouping by feature extension.
Remove unused 'Benchmark' project.
Merge all Simple.Web 'core' tests into the Simple.Web.Tests assembly.
Cleanup .gitignore.
Add SimpleWeb.Style.DotSettings R# v8 style settings file.
