+++ .gitignore
@@ -1,46 +1,51 @@
 obj
 bin
+build
+artifacts
 deploy
+docs
+nuget
+src/CommonAssemblyInfo.cs
+src/packages/*
+!src/packages/repositories.config
+
 *.csproj.user
 *.csproj.*.user
 *.suo
 *.cache
 *.nupkg
+
 ~$*
 *~
-*ReSharper*
+
 *.swp
-*.testsettings
-TestResults/*
 *.dbmdl
-Releases/2*
 *.pidb
 *.docstates
-test-results
-apply_config.bat
-NDependOut
-*.dotCover
+*.testsettings
 *_mm_cache.bin
-Simple.Data.sln.DotSettings.user
-Simple.Data/Simple.Data.idc
 .DS_Store
-mono-release-*
-Simple.Web.sln.DotSettings.user
-test
-test*ml
-build
-*.ncrunchproject
-Simple.Web.ncrunchsolution.user
-docs
-*.ndproj
+
+apply_config.bat
 nohup.out
-*.Tests.dll.xml
-*.DotSettings
-src/CommonAssemblyInfo.cs
 Gemfile.lock
+
+TestResults/*
+test
 results
-nuget
-src/packages/*
-!src/packages/repositories.config
-artifacts
-*.ncrunchsolution
+test-results
+test*ml
+*.Tests.dll.xml
+
+Releases/2*
+mono-release-*
+
+*ReSharper*
+*.DotSettings.*
+*.dotCover
+
+*.ncrunchsolution*
+*.ncrunchproject
+
+NDependOut
+*.ndproj
