--- src/Performance/Benchmark/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Fix" version="0.3.2.10" targetFramework="net45" />
-  <package id="Fix.AspNet" version="0.3.2.10" targetFramework="net45" />
-</packages>
