+++ src/Simple.Web.Tests/HandlerInfoTests.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Simple.Web.Tests
+﻿namespace Simple.Web.Tests
 {
+    using System;
     using Behaviors;
     using Hosting;
     using Xunit;
