+++ src/Simple.Web.Tests/LinkHelperTests.cs
@@ -1,7 +1,6 @@
 ﻿namespace Simple.Web.Tests
 {
     using System.Linq;
-    using Helpers;
     using Links;
     using Xunit;
