+++ src/Simple.Web.Tests/UriFromTypeTests.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Simple.Web.Tests
+﻿namespace Simple.Web.Tests
 {
     using Helpers;
     using Xunit;
